% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-uniform.R
\name{as_unisph_sample}
\alias{as_unisph_sample}
\title{Coercion method for \code{unisph_sample} objects}
\usage{
as_unisph_sample(x)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3}.}
}
\value{
An object of class \code{unisph_sample} and \code{sphere_sample}.
}
\description{
Coercion method for \code{unisph_sample} objects
}
\examples{
x <- runisph(100)
as_unisph_sample(x)
}
