% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-vmf.R
\name{as_vmf_sample}
\alias{as_vmf_sample}
\title{Coercion method for \code{vmf_sample} objects}
\usage{
as_vmf_sample(x)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3}.}
}
\value{
An object of class \code{vmf_sample} and \code{sphere_sample}.
}
\description{
Coercion method for \code{vmf_sample} objects
}
\examples{
x <- rvmf(100, c(1, 0, 0), 10)
as_vmf_sample(x)
}
