% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-watson.R
\name{as_watson_sample}
\alias{as_watson_sample}
\title{Coercion method for \code{watson_sample} objects}
\usage{
as_watson_sample(x)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3}.}
}
\value{
An object of class \code{watson_sample} and \code{sphere_sample}.
}
\description{
Coercion method for \code{watson_sample} objects
}
\examples{
x <- rwatson(100, c(1, 0, 0), 10)
as_watson_sample(x)
}
