% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-vmf.R
\name{fit.vmf_sample}
\alias{fit.vmf_sample}
\title{Fit a von Mises-Fisher distribution via MLE}
\usage{
\method{fit}{vmf_sample}(object, ...)
}
\arguments{
\item{object}{An object of class \code{vmf_sample} specifying a sample of
directions on the 3-sphere.}

\item{...}{Additional arguments. Currently unused; must be empty.}
}
\value{
A list containing the fitted parameters \code{mu} (mean direction) and
\code{kappa} (concentration parameter) of the von Mises-Fisher distribution.
}
\description{
Fit a von Mises-Fisher distribution via MLE
}
\examples{
x <- rvmf(100, c(1, 0, 0), 10)
fit(x)
}
