% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-watson.R
\name{fit.watson_sample}
\alias{fit.watson_sample}
\title{Fit a Watson distribution via MLE}
\usage{
\method{fit}{watson_sample}(object, ...)
}
\arguments{
\item{object}{An object of class \code{watson_sample} specifying a sample of
directions on the 3-sphere.}

\item{...}{Additional arguments. Currently unused; must be empty.}
}
\value{
A list containing the fitted parameters \code{mu} (mean axis) and \code{kappa}
(concentration parameter) of the Watson distribution.
}
\description{
Fit a Watson distribution via MLE
}
\examples{
x <- rwatson(100, c(1, 0, 0), 10)
fit(x)
}
