% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-vmf.R
\name{mean.vmf_sample}
\alias{mean.vmf_sample}
\title{Compute the mean of a \code{vmf_sample} object}
\usage{
\method{mean}{vmf_sample}(x, ...)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3}.}

\item{...}{Additional arguments. Not used.}
}
\value{
A numeric matrix of shape \eqn{1 \times 3} storing the mean of the
\code{vmf_sample} object.
}
\description{
Compute the mean of a \code{vmf_sample} object
}
\examples{
x <- rvmf(100, c(1, 0, 0), 10)
mean(x)
}
