% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-watson.R
\name{mean.watson_sample}
\alias{mean.watson_sample}
\title{Compute the mean of a \code{watson_sample} object}
\usage{
\method{mean}{watson_sample}(x, ...)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3}.}

\item{...}{Additional arguments. Not used.}
}
\value{
A numeric matrix of shape \eqn{1 \times 3} storing the mean of the
\code{watson_sample} object.
}
\description{
Compute the mean of a \code{watson_sample} object
}
\examples{
x <- rwatson(100, c(1, 0, 0), 10)
mean(as_watson_sample(x))
}
