% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sphere_sample}
\alias{plot.sphere_sample}
\title{Plot sphere sample}
\usage{
\method{plot}{sphere_sample}(x, opacity = 0.1, title = "3D Sphere Plot", ...)
}
\arguments{
\item{x}{An object of class \code{sphere_sample} specifying a sample of points on
a sphere. It is typically produced by functions like \code{rvmf()}, \code{rwatson()},
or similar functions that generate random samples on a sphere. One can also
add the \code{sphere_sample} class to a numeric matrix of shape \eqn{n \times 3}
where \eqn{n} is the number of samples and the columns represent the x, y,
z coordinates of the samples.}

\item{opacity}{A numeric value between 0 and 1 specifying the opacity of the
sphere surface. Defaults to \code{0.1}.}

\item{title}{A string specifying the title of the plot. Defaults to \code{"3D Sphere Plot"}.}

\item{...}{Additional arguments passed to \code{plotly::add_markers()} to
customize the appearance of the points in the plot, such as \code{color},
\code{size}, etc.}
}
\value{
A plotly object representing a 3D scatter plot of points on a sphere.
}
\description{
This function creates a 3D scatter plot of points on a sphere
using \code{plotly}. It visualizes a sample of points on a sphere, which can be
generated using various functions like \code{rvmf()}, \code{rwatson()}, or similar
functions that generate random samples on a sphere. The function allows
customization of the sphere's resolution, opacity, and title of the plot.
}
\examples{
x <- rwatson(100, c(1, 0, 0), 10)
as_watson_sample(x)
plot(x)
plot(x, color = I("darkblue"))
}
