% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-uniform.R
\name{Uniform}
\alias{Uniform}
\alias{dunisph}
\alias{punisph}
\alias{runisph}
\title{The Uniform distribution}
\usage{
dunisph(x, log = FALSE)

punisph(x)

runisph(n)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3} where \eqn{n} is the
number of samples and the columns represent the x, y, z coordinates
of the samples.}

\item{log}{A logical value indicating whether the log density should be
returned. Defaults to \code{false}.}

\item{n}{An integer value indicating the number of samples to generate.}
}
\value{
\itemize{
\item \code{dunisph} returns a numeric vector of length \eqn{n} containing the
density values of the Uniform distribution.
\item \code{punisph} returns a numeric vector of length \eqn{n} containing the
cumulative density values of the Uniform distribution.
\item \code{qunisph} returns a numeric matrix of shape \eqn{n \times 3} containing
the quantile values of the Uniform distribution.
\item \code{runisph} returns a numeric matrix of shape \eqn{n \times 3} containing
the random samples from the Uniform distribution.
}
}
\description{
The Uniform distribution
}
\examples{
n <- 100
spl <- runisph(n)
dunisph(spl)
punisph(spl)

}
