% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VonMisesFisher}
\alias{VonMisesFisher}
\alias{dvmf}
\alias{pvmf}
\alias{rvmf}
\title{The Von-Mises Fisher distribution}
\usage{
dvmf(x, mu, kappa, log = FALSE)

pvmf(x, mu, kappa)

rvmf(n, mu, kappa)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3} where \eqn{n} is the
number of samples and the columns represent the x, y, z coordinates
of the samples.}

\item{mu}{A numeric matrix of shape \eqn{1 \times 3} representing the mean
axis of the Von Mises-Fisher distribution.}

\item{kappa}{A positive numeric value representing the concentration
parameter of the Von Mises-Fisher distribution.}

\item{log}{A logical value indicating whether the log density should be
returned. Defaults to \code{false}.}

\item{n}{An integer value indicating the number of samples to generate.}
}
\value{
\itemize{
\item \code{dvmf} returns a numeric vector of length \eqn{n} containing the
density values of the Von Mises-Fisher distribution.
\item \code{pvmf} returns a numeric vector of length \eqn{n} containing the
cumulative density values of the Von Mises-Fisher distribution.
\item \code{qvmf} returns a numeric matrix of shape \eqn{n \times 3} containing
the quantile values of the Von Mises-Fisher distribution.
\item \code{rvmf} returns a numeric matrix of shape \eqn{n \times 3} containing
the random samples from the Von Mises-Fisher distribution.
}
}
\description{
The Von-Mises Fisher distribution
}
\examples{
mu <- c(1, 0, 0)
kappa <- 10
n <- 100
spl <- rvmf(n, mu, kappa)
dvmf(spl, mu, kappa)
pvmf(spl, mu, kappa)

}
