% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-watson.R
\name{Watson}
\alias{Watson}
\alias{dwatson}
\alias{pwatson}
\alias{rwatson}
\title{The Watson distribution}
\usage{
dwatson(x, mu, kappa, log = FALSE)

pwatson(x, mu, kappa)

rwatson(n, mu, kappa)
}
\arguments{
\item{x}{A numeric matrix of shape \eqn{n \times 3} where \eqn{n} is the
number of samples and the columns represent the x, y, z coordinates
of the samples.}

\item{mu}{A numeric matrix of shape \eqn{1 \times 3} representing the mean
axis of the Watson distribution.}

\item{kappa}{A positive numeric value representing the concentration
parameter of the Watson distribution.}

\item{log}{A logical value indicating whether the log density should be
returned. Defaults to \code{false}.}

\item{n}{An integer value indicating the number of samples to generate.}
}
\value{
\itemize{
\item \code{dwatson} returns a numeric vector of length \eqn{n} containing the
density values of the Watson distribution.
\item \code{pwatson} returns a numeric vector of length \eqn{n} containing the
cumulative density values of the Watson distribution.
\item \code{qwatson} returns a numeric matrix of shape \eqn{n \times 3} containing
the quantile values of the Watson distribution.
\item \code{rwatson} returns a numeric matrix of shape \eqn{n \times 3} containing
the random samples from the Watson distribution.
}
}
\description{
The Watson distribution
}
\examples{
mu <- c(1, 0, 0)
kappa <- 10
n <- 100
spl <- rwatson(n, mu, kappa)
dwatson(spl, mu, kappa)
pwatson(spl, mu, kappa)

}
