% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericaldistributions-package.R
\docType{package}
\name{sphericaldistributions-package}
\alias{sphericaldistributions}
\alias{sphericaldistributions-package}
\title{sphericaldistributions: What the Package Does (One Line, Title Case)}
\description{
This package aims to gather statistical distributions on the sphere, in order to allow a quick and easy way to fit your data to any distribution you like, but also find the best suiting distribution. We provide methods to compute the MLE, PDF, CDF, random number generation, and parameter estimation. We focused on making a lightweight package, with few dependencies. This package depends on the C++ Anima library \url{https://anima.readthedocs.io/en/latest/}, which is a library and set of python scripts for medical image processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LMJL-Alea/sphericaldistributions}
  \item Report bugs at \url{https://github.com/LMJL-Alea/sphericaldistributions/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Authors:
\itemize{
  \item Emmanuel Caruyer \email{emmanuel.caruyer@irisa.fr}
  \item Marie Poirier \email{marie.poirier@inria.fr}
}

Other contributors:
\itemize{
  \item Anima authors (https://anima.irisa.fr/) [contributor, copyright holder]
}

}
