% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericaldistributions-package.R
\docType{package}
\name{sphericaldistributions-package}
\alias{sphericaldistributions}
\alias{sphericaldistributions-package}
\title{sphericaldistributions: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LMJL-Alea/sphericaldistributions}
  \item Report bugs at \url{https://github.com/LMJL-Alea/sphericaldistributions/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
